finished parking place implementation

+++ frontend/src/app/components/admin-parking-place-selection-component/admin-parking-place-selection-component.component.ts
@@ -23,6 +23,7 @@ export class AdminParkingPlaceSelectionComponent implements OnInit {
 
   @Output() onDataChange: EventEmitter<any> = new EventEmitter<any>();
   @Output() onSpotDelete: EventEmitter<any> = new EventEmitter<any>();
+  @Output() onNewlyCreatedSpotDelete: EventEmitter<any> = new EventEmitter<any>();
 
   popupDescription = 'Do you really want to delete the parking place? Assigned available bikes will be unassigned!'
 
@@ -46,19 +47,21 @@ export class AdminParkingPlaceSelectionComponent implements OnInit {
   }
 
   fetchAssignedBike() {
-    this.spotService.fetchAssignedBike(this.spot.spot_id).subscribe({
-      next: (val) => {
-        console.log(val);
-        this.adminParkingPlaceAssignedBike = 'Assigned Bike: ' + val[0].bike_name;
-        this.bikeId = val[0].bike_id;
-        this.assignedBikeStatus = val[0].status;
-      },
-      error: (err) => {
-        if (err.status === 404) this.adminParkingPlaceAssignedBike = 'Assigned Bike: None';
-        else console.log("Error while fetching assigned bike: ", err)
+    if (this.spot.spot_id) {
+      this.spotService.fetchAssignedBike(this.spot.spot_id).subscribe({
+        next: (val) => {
+          console.log(val);
+          this.adminParkingPlaceAssignedBike = 'Assigned Bike: ' + val[0].bike_name;
+          this.bikeId = val[0].bike_id;
+          this.assignedBikeStatus = val[0].status;
+        },
+        error: (err) => {
+          if (err.status === 404) this.adminParkingPlaceAssignedBike = 'Assigned Bike: None';
+          else console.log("Error while fetching assigned bike: ", err)
 
-      }
-    })
+        }
+      })
+    } else this.adminParkingPlaceAssignedBike = 'Assigned Bike: None';
   }
 
 
@@ -73,23 +76,26 @@ export class AdminParkingPlaceSelectionComponent implements OnInit {
   }
 
   executeDeletionQuery() {
-    if(this.bikeId > 0) {
-      this.spotService.deleteParkingSpotWithBikeId(this.spot.spot_id, this.bikeId).subscribe({
-        next: (val) => {
-          this.onSpotDelete.emit();
-          console.log(val)
-        },
-        error: (err) => console.log("Error while deleting parking spot: ", err)
-      })
-    } else {
-      this.spotService.deleteParkingSpot(this.spot.spot_id).subscribe({
-        next: (val) => {
-          this.onSpotDelete.emit();
-          console.log(val)
-        },
-        error: (err) => console.log("Error while deleting parking spot: ", err)
-      })
+    if(this.spot.spot_id) {
+      if (this.bikeId > 0) {
+        this.spotService.deleteParkingSpotWithBikeId(this.spot.spot_id, this.bikeId).subscribe({
+          next: (val) => {
+            this.onSpotDelete.emit();
+            console.log(val)
+          },
+          error: (err) => console.log("Error while deleting parking spot: ", err)
+        })
+      } else {
+        this.spotService.deleteParkingSpot(this.spot.spot_id).subscribe({
+          next: (val) => {
+            this.onSpotDelete.emit();
+            console.log(val)
+          },
+          error: (err) => console.log("Error while deleting parking spot: ", err)
+        })
+      }
+    } else { // Newly created spot gets deleted
+      this.onNewlyCreatedSpotDelete.emit(this.spot.created_spot);
     }
-
   }
 }
